Allow taxnames not in ref, disallow genomes not in taxfile
Before, benchmarking threw an error when loading a taxfile with genomes that
did not exist in the reference. This is annoying, since we sometimes subset our
reference and would like to be able to use same tax file.
However, we should disallow taxfiles that do not specify taxonomy of all the
genomes in reference.