Fixup workflow Snakefile
Do not use an env file to activate the vamb conda environment using snakemake.
Instead, assume the user has an environment called "vamb" activate with the
correct packages.
This is kosher, since the installation instructions make sure such an env exist.

Also fix small bug in the parsing of the FASTQ file locations.