Parse BAM files in chunks
Current implementation uses too much memory in pycoverm when passed lots of BAM
files and threads. This is partly due to an inefficient design of pycoverm.
This change solves it on the Python side by batching the pycoverm calls and
saving temporary matrices to disk.

I'm currently working on improving the memory usage of pycoverm. If this is
sufficiently improved, AND I can verify its low memory usage in practical tests,
this commit will be reverted.