Add Black formatter
The only real change here is the addition of `.github/workflows/black.yml`.
This will test for each PR that the code is formatted according to the
[Black formatter](https://black.readthedocs.io/en/stable/).
All the remaining changes are simply from running Black on this repo.

To make your PRs pass, simply install `black` from pip, then call `black .`
when in the Vamb repo to format all your files.

This might be too invasive, but let's see. We can always revert it.

See also commits e6480ce and 5a0cc3a, the likes of which this PR is intended to
make obsolete.