Clustering: Adapt hyperparams to larger clusters
When investigating latent space on real assembled contigs, we see that the
clusters are more dispersed and wider compared to gold standard assemblies.
This is unsurprising, as real contigs are expected to be significantly more
noisy.
I conclude that the previous hyperparameters have been optimised for unrealistic
contigs, and easy gains could possibly be made by changing these.

Update some of the hyperparameters of the clustering algorithm so it works
better for larger, more realistic clusters.