Improve error message on mismatching refhash
Mismatching identifiers in FASTA and BAM files has caused a lot of issues for
our users. This is what the mechanism of refhashing was intended to detect
and prevent.
However, the error on a mismatching refhash is not that friendly, in that it
only shows the hash, but not the identifiers that caused the hash.
Improve the error message by showing the first mismatching identifier.