Add marker gene detection
Add the ability of Vamb to detect marker genes. This is done using pyrodigal
to predict genes in the contigs, and pyhmmer to find marker genes in the
predicted genes.

The marker genes are currently not used in Vamb, so future commits need to add
the ability of Vamb to enhance binning based on these marker genes.