Implement seed ordering
Before, the order in which seed contigs were picked was randomized, in order to
prevent bias arising from the order of the sequences in the FASTA files.
However, I am about to implement marker gene clustering, and review the effect
of contig length on clustering. In that context, picking contigs as seed which
we know a priory are more likely to be better seeds are important: We don't
want the clusterer to spend most its time discarding short and bad contigs as
potential seeds.
Hence, we pick the seeds in a precomputed order. In this commit, the order is
a combination of the contig length and the number of bacterial marker genes.
In the future, we might want to refine this order to include eukaryotes or such.
The idea is that the contigs which we are most confident should belong to a good
cluster gets "first pick" during clustering to see if they can form proper
clusters.