Add absolute abundance as VAE feature
The model now uses two distinct features of abundance: One tensor with absolute
abundances, and one with relative abundances.
This simplifies the code in that both features are always present no matter the
number of samples - even if the relative abundances are uninformative with one
sample.
It also extracts more information from the abundances, since only using the
relative abundances, as we did previously, removes a degree of freedom from each
contig's abundances.
In particular, this allows the network to distinguish the abundances of contigs
that are only present in a single sample, which is a surprising amount of them.

Unlike before, where we only used absolute abundances when given a single sample,
the code in this commit normalises abundances to zscores, which helps the model
learn when the loss used is SSE.

In the current implementation, the relative weight of absolute versus relative
abundances are that the absolute weighs 1/N and relative (N-1)/N, where N is the
number of samples. This is derived from the missing degree of freedom.

Benchmarking confirms a large improvement in performance with three samples, and
a medium improvement with 10 samples for the original CAMI2 datasets.
It may be worth experimenting with another loss weigh between abs/rel abundance.