Cache sampling of medoid
This commit adds a cache to the sampling of the medoid in clustering. This allows
the clusterer to fetch the results from a dict, if it's been previously computed.
This will get merged only if benchmarks have verified that clustering actually
runs faster - it's probable that so few medoids are sampled multiple times that
the cache wastes more time in overhead than it saves in computation.