Output .tsv files in actual TSV format
It turns out TSV has an actual specification, see
https://www.iana.org/assignments/media-types/text/tab-separated-values
and https://www.loc.gov/preservation/digital/formats/fdd/fdd000533.shtml#notes
However, Vamb did not follow this specification, as the header line was
missing. This is problematic, because external TSV parsers may not correctly
parse the output files. It's also just annoying for users.

In this commit, add a header line to all output .TSV files.
This is a breaking change, so let's get it in before v5.