Bugfix: Check for zero TNFs before projection
Before, when creating the dataloader, we raised an error if any contigs had
observed zero TNFs, since this would imply a sequence that was only composed of
ambiguous nucleotides, and thus an uninformative sequence (and almost certainly
also a mistake in the input).
However, since this check was in the data loader, the check happened on the
*projected* TNF space, and so the check for a sum of zero was meaningless.
Indeed, not only would this error not be triggered when an actual zero TNF seq
was observed, it would be falsely triggered in the very unlikely even that a
sequence's projected TNF summed to exactly zero.