Remove overfitting warning
In experiments, we've been unable to show that Vamb overfits, even with datasets
10% the size of our smallest test dataset. We're not sure why, but it's probably
because Vamb is extremely regularized already, and the regularization kicks
in only with a low number of contigs.

Vamb might still overfit on high quality assemblies like HiFi data, if there are
only a few thousand or few hundred contigs, but we cannot assess this until we
get good benchmark datasets of this quality.