Add stderr logger early in main
Because we remove the default logger on init, but only add the new logger when
run() is called, there is no logger available during the preprocessing steps,
e.g. argparsing.
This leads to warnings not being shown to the user, which is very bad from a
usability point of view.