Hide several less-used CLI options
Let's be real, what are the chances a Vamb user is going to experiment with tweaking
alpha or the latent space size?
It's nice to have these flags when testing, but we don't need to display them
in the help menu. All they do is confuse users and drown out the actually
important parameters users need to set.