Replace AEMB parser with generic TSV parser
In issue 355, a user requested to bring back the --jgi option, suc that users
can pass in precomputed abundances. That's completely reasonable, and it's not
great that the currently recommended way is to use the internal Python types
to create an Abundance object.

Instead of bringing back the JGI parser, I made a more generic TSV parser.
TSV is a simple, straightforward format also used elsewhere in Vamb. Also, since
it's easy to convert the output from --aemb to TSV format, I removed the --aemb
option.