Silence NumPy warning about float subnormals
The dependency pyhmmer is compiled with -funsafe-math-optimizations. Loading a
lib compiled with that flag will set some flags in the CPU for the whole
process, which changes float subnormal behaviour.
This is detected when loading NumPy, raising a warning.
I believe this is not an issue for Vamb, so I just silence the warning for now.
I'm not too happy with this "solution".