Rewrite reclustering (#340)
This large commit is a rewrite of reclustering, along with a variety of related 
fixes and changes:

* Add a new `taxonomy` module which includes the `ContigTaxonomy` previously
  stored in vambtools. Also include new types `PredictedTaxonomy` and `Taxonomy`
  to more clearly encapsulate the difference between the output of taxometer
  and that of e.g. mmseqs.
* Rewrite reclustering, such that it uses the `Markers` object, and parsemarkers
  integrates with reclustering better.
  The new version is more readable and extendable, is statically analyzable,
  and should be less brittle.
  It does not rely on the shell. 
* Remove dependency on pandas

Note that the changes in this commit has not been properly tested. That is
because following this commit, we will do a more thorough performance testing
of Vamb before release v5, and I expect that this process will cause a series
of fixes anyway.