Clear invalidated medoid cache on packing
Packing the indices of a ClusterGenerator shifts the meaning of the integers
in its cache, invalidating it.
I suppose they could be updated in the cache, but that'd be a complex piece of
code a clustering implementation which is already too complex for its own good.
Instead, clear the cache when packing.

This should happen a neglible amount of times compared to how often we already
clear the cache.