Also add bin prefix to AAE cluster names
In commit 905115, the bin files were renamed by adding a prefix in order to
resolve a file name conflict between the Z and Y space clusters.
However, this meant that the bin/cluster names now did not correspond to the
file names of the bins.
Instead, in this commit, rename the bins themselves.