Mask short contigs when loading TSV abundance
Previously, the TSV file was expected to only contain the retained contigs.
This is confusing for users, since they natually assume the TSV file should
contain the same contigs as the composition or the FASTA file.
So, also mask the TSV file with the composition metadata mask when loading it.